requirements.txt
fastapi
uvicorn
jinja2
google-generativeai
requests